+++ src/window.d.ts
@@ -1,3 +1,7 @@
+interface DataLayerObject extends Record<string, any> {
+  event: string;
+}
+
 declare interface Window {
-  dataLayer?: Array<Record<string, any>>;
+  dataLayer?: DataLayerObject[];
 }